changes to replies

+++ resources/assets/js/components/Reply.vue
@@ -18,8 +18,11 @@
 
             },
             destroy() {
-                axios.delete('/replies/'+this.attributes.id);
-                this.visible_element = false;
+                if (confirm("Are you sure")) {
+                    axios.delete('/replies/'+this.attributes.id);
+                    this.visible_element = false;
+                }
+
             }
